exceptions handled

+++ src/main/java/com/aimcodes/fashionBlog/services/serviceImpl/CategoryServiceImpl.java
@@ -3,6 +3,8 @@ package com.aimcodes.fashionBlog.services.serviceImpl;
 import com.aimcodes.fashionBlog.entities.Category;
 import com.aimcodes.fashionBlog.entities.User;
 import com.aimcodes.fashionBlog.enums.Role;
+import com.aimcodes.fashionBlog.exceptions.HandleNullException;
+import com.aimcodes.fashionBlog.exceptions.NoAccessException;
 import com.aimcodes.fashionBlog.pojos.ApiResponse;
 import com.aimcodes.fashionBlog.pojos.CategoryRequestDto;
 import com.aimcodes.fashionBlog.pojos.CategoryResponseDto;
@@ -30,55 +32,54 @@ public class CategoryServiceImpl implements CategoryService {
     public ResponseEntity<ApiResponse> createCategory(CategoryRequestDto request, HttpSession session) {
         User user = (User)session.getAttribute("currUser");
         if(user.getRole().equals(Role.valueOf("ADMIN"))){
-            Category category = new Category();
-            category.setName(request.getName().toLowerCase());
+            Category category = Category.builder().name(request.getName().toLowerCase()).build();
             categoryRepository.save(category);
 
-            CategoryResponseDto response = new CategoryResponseDto();
-            response.setName(category.getName());
-
+            CategoryResponseDto response = CategoryResponseDto.builder().name(request.getName()).build();
             return new ResponseEntity<>(responseManager.successfulRequest(response), HttpStatus.CREATED);
         }
-        return null;
+        throw new NoAccessException("Invalid user", "User is not an admin");
     }
 
     @Override
     public ResponseEntity<ApiResponse> updateCategory(CategoryRequestDto request, Long category_id, HttpSession session) {
         Category category = categoryRepository.findById(category_id).orElseThrow(null);
         User user = (User)session.getAttribute("currUser");
-        CategoryResponseDto response = new CategoryResponseDto();
-        if(user.getRole().equals(Role.valueOf("ADMIN"))){
+        if(category != null && user.getRole().equals(Role.valueOf("ADMIN"))){
             category.setName(request.getName());
             categoryRepository.save(category);
 
-            response.setName(category.getName());
-        }
-        return new ResponseEntity<>(responseManager.successfulRequest(response), HttpStatus.ACCEPTED);
+            CategoryResponseDto response = CategoryResponseDto.builder()
+                            .name(category.getName())
+                                    .build();
+            return new ResponseEntity<>(responseManager.successfulRequest(response), HttpStatus.ACCEPTED);
+        } else if (category == null)
+            throw new HandleNullException("Invalid category", "Category with id " + category_id + " does not exist in database");
+        throw new NoAccessException("Invalid user", "User is not an admin");
     }
 
     @Override
     public ResponseEntity<ApiResponse> deleteCategory(Long category_id, HttpSession session) {
         Category category = categoryRepository.findById(category_id).orElseThrow(null);
         User user = (User)session.getAttribute("currUser");
-        if(user.getRole().equals(Role.valueOf("ADMIN"))){
+        if(category != null && user.getRole().equals(Role.valueOf("ADMIN"))){
             categoryRepository.delete(category);
 
             return new ResponseEntity<>(responseManager.successfulRequest("Category deleted successful"), HttpStatus.ACCEPTED);
-        }
-        return null;
+        } else if (category == null)
+            throw new HandleNullException("Invalid category", "Category with id " + category_id + " does not exist in database");
+        throw new NoAccessException("Invalid user", "User is not an admin");
     }
 
     @Override
     public ResponseEntity<ApiResponse> view_all_Categories() {
         List<Category> allCategories = categoryRepository.findAll();
         List<CategoryResponseDto> responses = new ArrayList<>();
-
         allCategories.forEach(category -> {
-
-            CategoryResponseDto response = new CategoryResponseDto();
-            response.setName(category.getName());
+            CategoryResponseDto response = CategoryResponseDto.builder()
+                            .name(category.getName())
+                                    .build();
             responses.add(response);
-
         });
         return new ResponseEntity<>(responseManager.successfulRequest(responses), HttpStatus.FOUND);
     }
